requirements.txt
# ARIA 依赖

# 核心
pyyaml>=6.0

# 语音识别（本地 Whisper）
faster-whisper>=1.0.0

# 录音
sounddevice>=0.4.6
soundfile>=0.12.1

# 截图
mss>=9.0.0
Pillow>=10.0.0

# 意图解析
openai>=1.0.0          # 可选，provider=openai 时需要

# TTS
edge-tts>=6.1.0        # 免费，微软云

# OBS 控制（Replay Buffer）
obsws-python>=1.7.0    # 可选，有 OBS 才需要

# 音频播放
playsound>=1.3.0       # 可选备选，也可用系统命令

# numpy（录音处理用）
numpy>=1.24.0
